# [hail] Fix `mt.entry.take` performance regression