--- a/hail/src/main/scala/is/hail/sparkextras/ContextRDD.scala
+++ b/hail/src/main/scala/is/hail/sparkextras/ContextRDD.scala
@@ -388,7 +388,7 @@ class ContextRDD[C <: AutoCloseable, T: ClassTag](
       case None =>
         val PCSubsetOffset(idx, nTake, _) =
           incrementalPCSubsetOffset(n, 0 until getNumPartitions)(
-            runJob(getIteratorSize, _)
+            runJob(getIteratorSizeWithMaxN(n), _)
           )
         idx -> nTake
     }